Create currently reading functionality + fix bugs
If user clicks on a library book, that library
book would be brought to the front and have
the text CURRENTLY READING at the top.

However, this lead to difficulties with previous
commits that had the finish button and the
heart icon stay active when pressed. In the end,
I scraped that idea for this idea that I believe
has more functionality